refactor: roadmap 이름 변경
